0921
0921 코드 정리 ,  Die 구현

+++ Assets/Player/Player.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class Player : MonoBehaviour
 {
-    public int playerHp = 10;
+    public float playerHp = 10;
     public float playerAttack = 1f;
 }
